Add a GH action to manage images provided in the report
If images are provided instead of "Game Display Settings" (meaning the section is empty or "_No response_" or only contains image links), then we will upload the images to our OCR API to parse them and extract the game configuration.

This adds a comment and a new label to the report to identify it as having been updated by OCR generated content that should be reviewed by the author. Once the author reviews the report content and saves the issue body again, then the next run of this action will remove that label and the comment.

Add some description to the game report template to indicate that it supports using AI to parse data from game screenshots